Make ui modular and extensible
 - Updated Angular to 1.5.3
 - Removed Angular UI Bootstrap, to make the Angular update possible without
   updating Bootstrap
 - Moved every view into its own module
 - Added concatenation of the module js and css assets to the webserver
 - Created directives/components for the various visual components and trimmed
   down the controllers
 - Removed unecessary ngRoute
 - Removed classpath section in details (it is included in environment)
 - Moved the counters and gauges to a separated metrics view
 - Added target="_blank" for the logfile
 - Added information about version, loaded ui modules to the about page
 - Extracted the activiti view to a separate maven artifact

With this commit it is now possible to add your own ui modules to the servers
classpath and they will show up in the ui.
Every module needs an "entry" script file in the classpath located under
`META-INF/spring-boot-admin-server-ui/<modulename>/module.js`.
The module is be implemented as angular module and needs to be added to
the global `uiModules` list.
The spring-boot-admin-server-ui-activiti module is a simple sample for
adding a application view but it is also possible to add main-views.
Please have a look at the events module in spring-boot-admin-server-ui/modules.

Please note that the internals of the ui can change anytime.