Don't ignore events from child contexts
With the fix for #260 we ignored all events from child contexts which turns out
to be wrong since the "main" application context is, in case spring cloud config
is used, is a child  of the bootstrap context. So we just act on the events from
WebApplicationContexts. After they're fired up the needed port information is
available.

fixes #277