Improve the client configuration for usage in war-file deployments
With this commit two new properties were added to ease the client
configuration when used in servlet containers:
`spring.boot.admin.client.management-base-url` and
`spring.boot.admin.client.service-base-url`. Both properties can be used
to specify protocol, host and port for the urls. The paths will still be
guessed and appended to the base-urls. Additionally the servlet context
path is not read from the properties but from the ServletContext itself.

closes #488