Invalidate snapshots after OptimistickLockingException
When an OptimisticLockingException occurs we need to explicitly
invalidate the snapshot for this particular entity and eagerly update.
Otherwise it is most likely that retrying the operation will result in
the same outcome as it is not guaranteed that the events updating the
snapshots will be consumed in between.

fixes #1059